requirements.txt
ace_tools==0.0
acres==0.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiohttp-retry==2.9.1
aiosignal==1.4.0
alembic==1.17.0
altair==5.5.0
amqp==5.3.1
annotated-doc==0.0.3
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
appdirs==1.4.4
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
asyncssh==2.22.0
atpublic==7.0.0
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.5
billiard==4.2.4
bleach==6.2.0
blinker==1.9.0
cachetools==6.2.1
celery==5.6.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
ci-info==0.3.0
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpickle==3.1.2
colorama==0.4.6
comm==0.2.3
configobj==5.0.9
configparser==7.2.0
contourpy==1.3.3
cryptography==46.0.1
curl_cffi==0.13.0
cycler==0.12.1
databricks-sdk==0.76.0
dataclasses-json==0.6.7
datasets==4.4.2
debugpy==1.8.16
decorator==5.2.1
defusedxml==0.7.1
dictdiffer==0.9.0
dill==0.4.0
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
docker==7.1.0
dpath==2.2.0
dulwich==0.25.0
dvc==3.65.0
dvc-data==3.17.0
dvc-http==2.32.0
dvc-objects==5.2.0
dvc-render==1.0.2
dvc-studio-client==0.22.0
dvc-task==0.40.2
dvclive==3.49.0
entrypoints==0.4
etelemetry==0.3.1
evaluate==0.4.6
exceptiongroup==1.3.1
executing==2.2.1
faiss-cpu==1.12.0
fastapi==0.121.0
fastjsonschema==2.21.2
filelock==3.20.0
fitz==0.0.1.dev2
Flask==3.1.2
flask-cors==6.0.2
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
flatten-dict==0.4.2
flufl.lock==8.2.0
fonttools==4.59.2
fqdn==1.5.1
frozendict==2.4.6
frozenlist==1.8.0
fsspec==2025.9.0
funcy==2.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.45.0
grandalf==0.8
graphene==3.4.3
graphql-core==3.2.7
graphql-relay==3.2.0
greenlet==3.2.4
gto==1.9.0
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.3
huey==2.5.5
huggingface-hub==0.35.3
hydra-core==1.3.2
idna==3.10
imbalanced-learn==0.14.0
importlib_metadata==8.7.1
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
isoduration==20.11.0
iterative-telemetry==0.0.10
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
json5==0.12.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.7
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
kombu==5.6.1
langchain==1.0.0
langchain-classic==1.0.0
langchain-community==0.4
langchain-core==1.0.0
langchain-huggingface==1.0.0
langchain-text-splitters==1.0.0
langgraph==1.0.0
langgraph-checkpoint==2.1.2
langgraph-prebuilt==1.0.0
langgraph-sdk==0.2.9
langsmith==0.4.37
lark==1.2.2
lightgbm==4.6.0
looseversion==1.3.0
lxml==6.0.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.4
mlflow==3.8.1
mlflow-skinny==3.8.1
mlflow-tracing==3.8.1
multidict==6.7.0
multiprocess==0.70.18
multitasking==0.0.12
mypy_extensions==1.1.0
narwhals==2.8.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
nltk==3.9.2
notebook==7.4.5
notebook_shim==0.2.4
numpy==2.3.3
nvidia-ml-py==13.590.44
omegaconf==2.3.0
opentelemetry-api==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.3
ormsgpack==1.11.0
packaging==25.0
pandas==2.3.2
pandocfilters==1.5.1
parso==0.8.5
pathlib==1.0.1
pathspec==0.12.1
peewee==3.18.2
pexpect==4.9.0
pillow==11.3.0
pipenv==2025.0.4
platformdirs==4.4.0
prometheus_client==0.22.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.32.1
prov==2.1.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==1.30
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
pydeck==0.9.1
pydot==4.0.1
pygit2==1.19.0
Pygments==2.19.2
pygtrie==2.5.0
PyMySQL==1.1.2
pynvml==13.0.1
pyparsing==3.2.4
pypdf==6.1.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-json-logger==3.3.0
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
pyzmq==27.1.0
rdflib==7.2.1
referencing==0.36.2
regex==2025.11.3
reportlab==4.4.4
requests==2.32.5
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rpds-py==0.27.1
rsa==4.9.1
ruamel.yaml==0.18.17
ruamel.yaml.clib==0.2.15
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.2
scmrepo==3.6.0
seaborn==0.13.2
semver==3.0.4
Send2Trash==1.8.3
sentencepiece==0.2.1
setuptools==80.9.0
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.8.0
simplejson==3.20.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
sqlparse==0.5.5
sqltrie==0.11.2
stack-data==0.6.3
starlette==0.49.3
streamlit==1.50.0
tabulate==0.9.0
tenacity==9.1.2
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.22.1
toml==0.10.2
tomlkit==0.13.3
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
traits==7.0.2
transformers==4.57.3
typer==0.20.1
types-python-dateutil==2.9.0.20250822
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.38.0
vine==5.1.0
virtualenv==20.35.3
voluptuous==0.16.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.15
wordcloud==1.9.4
xgboost==3.1.2
xxhash==3.6.0
yarl==1.22.0
yfinance==0.2.65
zc.lockfile==4.0
zipp==3.23.0
zstandard==0.25.0
